refactor: Extract focused modules to comply with 500-line limit
Refactored files exceeding Pylint max-module-lines=500 by extracting
focused modules instead of adding ignore comments:

New utility modules:
- src/core/violation_utils.py - Shared violation line helpers
- src/linters/dry/single_statement_detector.py - Python AST patterns
- src/linters/dry/typescript_statement_detector.py - TS tree-sitter patterns
- src/linters/magic_numbers/typescript_ignore_checker.py - TS ignore checking

Refactored files:
- magic_numbers/linter.py: 516 → 461 lines (11% reduction)
- dry/python_analyzer.py: 684 → 281 lines (59% reduction)
- dry/typescript_analyzer.py: 622 → 274 lines (56% reduction)

Also updated print_statements/linter.py to use shared violation_utils.
