+++ src/linters/dry/typescript_analyzer.py
@@ -30,7 +30,7 @@ SRP Exception: TypeScriptDuplicateAnalyzer has 20 methods and 324 lines (exceeds
     responsibility: accurately detecting duplicate TypeScript/JavaScript code while minimizing false positives.
 """
 
-from collections.abc import Generator, Iterable
+from collections.abc import Iterable
 from pathlib import Path
 
 from src.analyzers.typescript_base import TREE_SITTER_AVAILABLE
@@ -39,6 +39,7 @@ from .base_token_analyzer import BaseTokenAnalyzer
 from .block_filter import BlockFilterRegistry, create_default_registry
 from .cache import CodeBlock
 from .config import DRYConfig
+from .typescript_statement_detector import TypeScriptStatementDetector
 
 if TREE_SITTER_AVAILABLE:
     from tree_sitter import Node
@@ -61,6 +62,7 @@ class TypeScriptDuplicateAnalyzer(BaseTokenAnalyzer):  # thailint: ignore[srp.vi
         """
         super().__init__()
         self._filter_registry = filter_registry or create_default_registry()
+        self._statement_detector = TypeScriptStatementDetector()
 
     def analyze(self, file_path: Path, content: str, config: DRYConfig) -> list[CodeBlock]:
         """Analyze TypeScript/JavaScript file for duplicate code blocks.
@@ -88,8 +90,8 @@ class TypeScriptDuplicateAnalyzer(BaseTokenAnalyzer):  # thailint: ignore[srp.vi
         valid_windows = (
             (hash_val, start_line, end_line, snippet)
             for hash_val, start_line, end_line, snippet in windows
-            if self._should_include_block(content, start_line, end_line)
-            and not self._is_single_statement_in_source(content, start_line, end_line)
+            if self._statement_detector.should_include_block(content, start_line, end_line)
+            and not self._statement_detector.is_single_statement(content, start_line, end_line)
         )
         return self._build_blocks(valid_windows, file_path, content)
 
@@ -269,354 +271,3 @@ class TypeScriptDuplicateAnalyzer(BaseTokenAnalyzer):  # thailint: ignore[srp.vi
             hashes.append((hash_val, start_line, end_line, snippet))
 
         return hashes
-
-    def _should_include_block(self, content: str, start_line: int, end_line: int) -> bool:
-        """Filter out blocks that overlap with interface/type definitions.
-
-        Args:
-            content: File content
-            start_line: Block start line
-            end_line: Block end line
-
-        Returns:
-            False if block overlaps interface definition, True otherwise
-        """
-        interface_ranges = self._find_interface_ranges(content)
-        return not self._overlaps_interface(start_line, end_line, interface_ranges)
-
-    def _is_single_statement_in_source(self, content: str, start_line: int, end_line: int) -> bool:
-        """Check if a line range in the original source is a single logical statement.
-
-        Uses tree-sitter AST analysis to detect patterns like:
-        - Decorators (@Component(...))
-        - Function call arguments
-        - Object literal properties
-        - Class field definitions
-        - Type assertions
-        - Chained method calls (single expression)
-
-        Args:
-            content: TypeScript source code
-            start_line: Starting line number (1-indexed)
-            end_line: Ending line number (1-indexed)
-
-        Returns:
-            True if this range represents a single logical statement/expression
-        """
-        if not TREE_SITTER_AVAILABLE:
-            return False
-
-        from src.analyzers.typescript_base import TypeScriptBaseAnalyzer
-
-        analyzer = TypeScriptBaseAnalyzer()
-        root = analyzer.parse_typescript(content)
-        if not root:
-            return False
-
-        return self._check_overlapping_nodes(root, start_line, end_line)
-
-    def _check_overlapping_nodes(self, root: Node, start_line: int, end_line: int) -> bool:
-        """Check if any AST node overlaps and matches single-statement pattern.
-
-        Args:
-            root: Root tree-sitter node
-            start_line: Starting line number (1-indexed)
-            end_line: Ending line number (1-indexed)
-
-        Returns:
-            True if any node matches single-statement pattern
-        """
-        # Convert to 0-indexed for tree-sitter
-        ts_start = start_line - 1
-        ts_end = end_line - 1
-
-        for node in self._walk_nodes(root):
-            if self._node_overlaps_and_matches(node, ts_start, ts_end):
-                return True
-        return False
-
-    def _walk_nodes(self, node: Node) -> Generator[Node, None, None]:
-        """Generator to walk all nodes in tree.
-
-        Args:
-            node: Starting node
-
-        Yields:
-            All nodes in tree
-        """
-        yield node
-        for child in node.children:
-            yield from self._walk_nodes(child)
-
-    def _node_overlaps_and_matches(self, node: Node, ts_start: int, ts_end: int) -> bool:
-        """Check if node overlaps with range and matches single-statement pattern.
-
-        Args:
-            node: Tree-sitter node
-            ts_start: Starting line (0-indexed)
-            ts_end: Ending line (0-indexed)
-
-        Returns:
-            True if node overlaps and matches pattern
-        """
-        node_start = node.start_point[0]
-        node_end = node.end_point[0]
-
-        # Check if ranges overlap
-        overlaps = not (node_end < ts_start or node_start > ts_end)
-        if not overlaps:
-            return False
-
-        return self._is_single_statement_pattern(node, ts_start, ts_end)
-
-    def _matches_simple_container_pattern(self, node: Node, contains: bool) -> bool:
-        """Check if node is a simple container pattern (decorator, object, etc.).
-
-        Args:
-            node: AST node to check
-            contains: Whether node contains the range
-
-        Returns:
-            True if node matches simple container pattern
-        """
-        simple_types = (
-            "decorator",
-            "object",
-            "member_expression",
-            "as_expression",
-            "array_pattern",
-        )
-        return node.type in simple_types and contains
-
-    def _matches_call_expression_pattern(
-        self, node: Node, ts_start: int, ts_end: int, contains: bool
-    ) -> bool:
-        """Check if node is a call expression pattern.
-
-        Args:
-            node: AST node to check
-            ts_start: Starting line (0-indexed)
-            ts_end: Ending line (0-indexed)
-            contains: Whether node contains the range
-
-        Returns:
-            True if node matches call expression pattern
-        """
-        if node.type != "call_expression":
-            return False
-
-        # Check if this is a multi-line call containing the range
-        node_start = node.start_point[0]
-        node_end = node.end_point[0]
-        is_multiline = node_start < node_end
-        if is_multiline and node_start <= ts_start <= node_end:
-            return True
-
-        return contains
-
-    def _matches_declaration_pattern(self, node: Node, contains: bool) -> bool:
-        """Check if node is a lexical declaration pattern.
-
-        Args:
-            node: AST node to check
-            contains: Whether node contains the range
-
-        Returns:
-            True if node matches declaration pattern (excluding function bodies)
-        """
-        if node.type != "lexical_declaration" or not contains:
-            return False
-
-        # Only filter if simple value assignment, NOT a function body
-        if self._contains_function_body(node):
-            return False
-
-        return True
-
-    def _matches_jsx_pattern(self, node: Node, contains: bool) -> bool:
-        """Check if node is a JSX element pattern.
-
-        Args:
-            node: AST node to check
-            contains: Whether node contains the range
-
-        Returns:
-            True if node matches JSX pattern
-        """
-        jsx_types = ("jsx_opening_element", "jsx_self_closing_element")
-        return node.type in jsx_types and contains
-
-    def _matches_class_body_pattern(self, node: Node, ts_start: int, ts_end: int) -> bool:
-        """Check if node is a class body field definition pattern.
-
-        Args:
-            node: AST node to check
-            ts_start: Starting line (0-indexed)
-            ts_end: Ending line (0-indexed)
-
-        Returns:
-            True if node is class body with field definitions
-        """
-        if node.type != "class_body":
-            return False
-
-        return self._is_in_class_field_area(node, ts_start, ts_end)
-
-    def _is_single_statement_pattern(self, node: Node, ts_start: int, ts_end: int) -> bool:
-        """Check if an AST node represents a single-statement pattern to filter.
-
-        Delegates to specialized pattern matchers for different AST node categories.
-
-        Args:
-            node: AST node that overlaps with the line range
-            ts_start: Starting line number (0-indexed)
-            ts_end: Ending line number (0-indexed)
-
-        Returns:
-            True if this node represents a single logical statement pattern
-        """
-        node_start = node.start_point[0]
-        node_end = node.end_point[0]
-        contains = (node_start <= ts_start) and (node_end >= ts_end)
-
-        # Check pattern categories using specialized helpers - use list for any()
-        matchers = [
-            self._matches_simple_container_pattern(node, contains),
-            self._matches_call_expression_pattern(node, ts_start, ts_end, contains),
-            self._matches_declaration_pattern(node, contains),
-            self._matches_jsx_pattern(node, contains),
-            self._matches_class_body_pattern(node, ts_start, ts_end),
-        ]
-        return any(matchers)
-
-    def _contains_function_body(self, node: Node) -> bool:
-        """Check if node contains an arrow function or function expression.
-
-        Args:
-            node: Node to check
-
-        Returns:
-            True if node contains a function with a body
-        """
-        for child in node.children:
-            if child.type in ("arrow_function", "function", "function_expression"):
-                return True
-            if self._contains_function_body(child):
-                return True
-        return False
-
-    def _find_first_method_line(self, class_body: Node) -> int | None:
-        """Find line number of first method in class body.
-
-        Args:
-            class_body: Class body node
-
-        Returns:
-            Line number of first method or None if no methods
-        """
-        for child in class_body.children:
-            if child.type in ("method_definition", "function_declaration"):
-                return child.start_point[0]
-        return None
-
-    def _is_in_class_field_area(self, class_body: Node, ts_start: int, ts_end: int) -> bool:
-        """Check if range is in class field definition area (before methods).
-
-        Args:
-            class_body: Class body node
-            ts_start: Starting line (0-indexed)
-            ts_end: Ending line (0-indexed)
-
-        Returns:
-            True if range is in field area
-        """
-        first_method_line = self._find_first_method_line(class_body)
-        class_start = class_body.start_point[0]
-        class_end = class_body.end_point[0]
-
-        # No methods: check if range is in class body
-        if first_method_line is None:
-            return class_start <= ts_start and class_end >= ts_end
-
-        # Has methods: check if range is before first method
-        return class_start <= ts_start and ts_end < first_method_line
-
-    def _find_interface_ranges(self, content: str) -> list[tuple[int, int]]:
-        """Find line ranges of interface/type definitions.
-
-        Args:
-            content: File content
-
-        Returns:
-            List of (start_line, end_line) tuples for interface blocks
-        """
-        ranges: list[tuple[int, int]] = []
-        lines = content.split("\n")
-        state = {"in_interface": False, "start_line": 0, "brace_count": 0}
-
-        for i, line in enumerate(lines, start=1):
-            stripped = line.strip()
-            self._process_line_for_interface(stripped, i, state, ranges)
-
-        return ranges
-
-    def _process_line_for_interface(
-        self, stripped: str, line_num: int, state: dict, ranges: list[tuple[int, int]]
-    ) -> None:
-        """Process single line for interface detection.
-
-        Args:
-            stripped: Stripped line content
-            line_num: Line number
-            state: Tracking state (in_interface, start_line, brace_count)
-            ranges: Accumulated interface ranges
-        """
-        if self._is_interface_start(stripped):
-            self._handle_interface_start(stripped, line_num, state, ranges)
-            return
-
-        if state["in_interface"]:
-            self._handle_interface_continuation(stripped, line_num, state, ranges)
-
-    def _is_interface_start(self, stripped: str) -> bool:
-        """Check if line starts interface/type definition."""
-        return stripped.startswith(("interface ", "type ")) and "{" in stripped
-
-    def _handle_interface_start(
-        self, stripped: str, line_num: int, state: dict, ranges: list[tuple[int, int]]
-    ) -> None:
-        """Handle start of interface definition."""
-        state["in_interface"] = True
-        state["start_line"] = line_num
-        state["brace_count"] = stripped.count("{") - stripped.count("}")
-
-        if state["brace_count"] == 0:  # Single-line interface
-            ranges.append((line_num, line_num))
-            state["in_interface"] = False
-
-    def _handle_interface_continuation(
-        self, stripped: str, line_num: int, state: dict, ranges: list[tuple[int, int]]
-    ) -> None:
-        """Handle continuation of interface definition."""
-        state["brace_count"] += stripped.count("{") - stripped.count("}")
-        if state["brace_count"] == 0:
-            ranges.append((state["start_line"], line_num))
-            state["in_interface"] = False
-
-    def _overlaps_interface(
-        self, start: int, end: int, interface_ranges: list[tuple[int, int]]
-    ) -> bool:
-        """Check if block overlaps with any interface range.
-
-        Args:
-            start: Block start line
-            end: Block end line
-            interface_ranges: List of interface definition ranges
-
-        Returns:
-            True if block overlaps with an interface
-        """
-        for if_start, if_end in interface_ranges:
-            if start <= if_end and end >= if_start:
-                return True
-        return False
