+++ src/linters/magic_numbers/linter.py
@@ -29,12 +29,14 @@ from pathlib import Path
 from src.core.base import BaseLintContext, MultiLanguageLintRule
 from src.core.linter_utils import load_linter_config
 from src.core.types import Violation
+from src.core.violation_utils import get_violation_line, has_python_noqa
 from src.linter_config.ignore import IgnoreDirectiveParser
 
 from .config import MagicNumberConfig
 from .context_analyzer import ContextAnalyzer
 from .python_analyzer import PythonMagicNumberAnalyzer
 from .typescript_analyzer import TypeScriptMagicNumberAnalyzer
+from .typescript_ignore_checker import TypeScriptIgnoreChecker
 from .violation_builder import ViolationBuilder
 
 
@@ -46,6 +48,7 @@ class MagicNumberRule(MultiLanguageLintRule):  # thailint: ignore[srp]
         self._ignore_parser = IgnoreDirectiveParser()
         self._violation_builder = ViolationBuilder(self.rule_id)
         self._context_analyzer = ContextAnalyzer()
+        self._typescript_ignore_checker = TypeScriptIgnoreChecker()
 
     @property
     def rule_id(self) -> str:
@@ -282,28 +285,17 @@ class MagicNumberRule(MultiLanguageLintRule):  # thailint: ignore[srp]
         Returns:
             True if line has generic ignore directive
         """
-        line_text = self._get_violation_line(violation, context)
+        line_text = get_violation_line(violation, context)
         if line_text is None:
             return False
 
         return self._has_generic_ignore_directive(line_text)
 
-    def _get_violation_line(self, violation: Violation, context: BaseLintContext) -> str | None:
-        """Get the line text for a violation."""
-        if not context.file_content:
-            return None
-
-        lines = context.file_content.splitlines()
-        if violation.line <= 0 or violation.line > len(lines):
-            return None
-
-        return lines[violation.line - 1].lower()
-
     def _has_generic_ignore_directive(self, line_text: str) -> bool:
         """Check if line has generic ignore directive."""
         if self._has_generic_thailint_ignore(line_text):
             return True
-        return self._has_noqa_directive(line_text)
+        return has_python_noqa(line_text)
 
     def _has_generic_thailint_ignore(self, line_text: str) -> bool:
         """Check for generic thailint: ignore (no brackets)."""
@@ -311,10 +303,6 @@ class MagicNumberRule(MultiLanguageLintRule):  # thailint: ignore[srp]
             return False
         after_ignore = line_text.split("# thailint: ignore")[1].split("#")[0]
         return "[" not in after_ignore
-
-    def _has_noqa_directive(self, line_text: str) -> bool:
-        """Check for noqa-style comments."""
-        return "# noqa" in line_text
 
     def _check_typescript(
         self, context: BaseLintContext, config: MagicNumberConfig
@@ -466,51 +454,4 @@ class MagicNumberRule(MultiLanguageLintRule):  # thailint: ignore[srp]
         Returns:
             True if should ignore
         """
-        # Check standard ignore directives
-        if self._ignore_parser.should_ignore_violation(violation, context.file_content or ""):
-            return True
-
-        # Check TypeScript-style comments
-        return self._check_typescript_ignore(violation, context)
-
-    def _check_typescript_ignore(self, violation: Violation, context: BaseLintContext) -> bool:
-        """Check for TypeScript-style ignore directives.
-
-        Args:
-            violation: Violation to check
-            context: Lint context
-
-        Returns:
-            True if line has ignore directive
-        """
-        line_text = self._get_violation_line(violation, context)
-        if line_text is None:
-            return False
-
-        # Check for // thailint: ignore or // noqa
-        return self._has_typescript_ignore_directive(line_text)
-
-    def _has_typescript_ignore_directive(self, line_text: str) -> bool:
-        """Check if line has TypeScript-style ignore directive.
-
-        Args:
-            line_text: Line text to check
-
-        Returns:
-            True if has ignore directive
-        """
-        # Check for // thailint: ignore[magic-numbers]
-        if "// thailint: ignore[magic-numbers]" in line_text:
-            return True
-
-        # Check for // thailint: ignore (generic)
-        if "// thailint: ignore" in line_text:
-            after_ignore = line_text.split("// thailint: ignore")[1].split("//")[0]
-            if "[" not in after_ignore:
-                return True
-
-        # Check for // noqa
-        if "// noqa" in line_text:
-            return True
-
-        return False
+        return self._typescript_ignore_checker.should_ignore(violation, context)
